requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-rest-passwordreset==1.3.0
django-solo==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
